# public/openapi.json
{"openapi":"3.0.0","info":{"title":"MileApp Fullstack Test","version":"v0.0.1","description":"## Introduction\nThis documentation describes the **Mock API** built for the *MileApp Fullstack Test*.\nThe API simulates authentication and task management endpoints without using a real database.\nIt is designed to demonstrate backend API structure, response standards, and frontend integration readiness.\n\nUse this reference to understand how to:\n- Authenticate using the `\/api\/login` endpoint.\n- Manage tasks through standard CRUD routes under `\/api\/tasks`.\n\nAll endpoints return mock data in **JSON** format and require **Bearer token authentication** for protected routes.\n\n---\n\n## HTTP Methods\n| Method | Description |\n|---------|-------------|\n| **GET** | Retrieve one or more resources |\n| **POST** | Create a new resource |\n| **PUT** | Update an existing resource |\n| **DELETE** | Remove a resource |\n\n---\n\n## HTTP Status Codes\n| Code | Description |\n|------|-------------|\n| **200** | OK \u2014 Request succeeded |\n| **201** | Created \u2014 Resource created successfully |\n| **204** | No Content \u2014 Request succeeded, no body returned |\n| **400** | Bad Request \u2014 The request was malformed |\n| **401** | Unauthorized \u2014 Missing or invalid Bearer token |\n| **403** | Forbidden \u2014 Authenticated but not permitted |\n| **404** | Not Found \u2014 The resource does not exist |\n| **422** | Unprocessable Entity \u2014 Validation failed |\n| **500** | Internal Server Error \u2014 Unexpected server failure |\n\n---\n\n### Notes\n- All timestamps use **ISO 8601** format.\n- Example data is generated statically for demonstration.\n- You can download the full OpenAPI specification at the top of this page.","termsOfService":"","license":null,"contact":[],"x-logo":{"url":"","altText":"MileApp Fullstack Test","backgroundColor":""}},"components":{"securitySchemes":{"BearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":[]},"servers":[{"url":"http:\/\/mileapp-api.test","description":"Documentation generator server."},{"url":"http:\/\/test.example.com","description":"Test server."}],"paths":{"\/api\/login":{"post":{"tags":["Auth"],"operationId":"Login","description":"","requestBody":{"description":"","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string","example":"user@example.com","description":"Email."},"password":{"type":"string","example":"secret123","description":"Password."}},"example":{"email":"user@example.com","password":"secret123"}}}}},"parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"message":"Mocked login successful","access_token":"FskV2W4bNq6Vckz4itsuXXRq1AT6rucRlnIS3mXM11bhlMYL6bbdwZPgDLfg","token_type":"Bearer"}}}}},"422":{"description":"error","content":{"application\/json":{"schema":{"type":"object","example":{"message":"The given data was invalid.","errors":{"email":["The email field is required."],"password":["The password field is required."]}}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/mileapp-api.test\/api\/login\" \n    -d \"email\"=\"user@example.com\" \\\n    -d \"password\"=\"secret123\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/mileapp-api.test\/api\/login\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"email\": \"user@example.com\",\n    \"password\": \"secret123\",\n})\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"email\" =>  \"user@example.com\",\n    \"password\" =>  \"secret123\",\n];\n\n\n$ch = curl_init(\"http:\/\/mileapp-api.test\/api\/login\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_POSTFIELDS, $body);\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/tasks":{"get":{"security":[{"BearerAuth":[]}],"tags":["Tasks"],"operationId":"List tasks","description":"","parameters":[{"in":"query","name":"page","description":"Current page.","required":false,"schema":{"type":"integer","example":1}},{"in":"query","name":"per_page","description":"Items per page.","required":false,"schema":{"type":"integer","example":10}},{"in":"query","name":"filter[is_completed]","description":"Filter completed.","required":false,"schema":{"type":"boolean","example":false}},{"in":"query","name":"sort_by","description":"Sort by.","required":false,"schema":{"type":"string","example":"created_at"}},{"in":"query","name":"sort_order","description":"Sort order.","required":false,"schema":{"type":"string","example":"desc"}}],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"data":[{"id":1,"title":"Design login page","description":"Create a responsive login form with validation and error handling.","due_date":"2025-11-10","is_completed":false,"completed_at":"2025-11-02 12:00:00","created_at":"2025-11-01 09:30:00","updated_at":"2025-11-03 14:45:00"},{"id":2,"title":"Implement task CRUD API","description":"Setup endpoints for create, read, update, and delete using mock data.","due_date":"2025-11-12","is_completed":false,"completed_at":"2025-11-02 12:00:00","created_at":"2025-11-02 10:15:00","updated_at":"2025-11-04 16:20:00"},{"id":3,"title":"Integrate frontend with backend","description":"Connect Vue frontend with Laravel mock API using Axios.","due_date":"2025-11-14","is_completed":false,"completed_at":"2025-11-02 12:00:00","created_at":"2025-11-03 11:00:00","updated_at":"2025-11-05 09:40:00"},{"id":4,"title":"Add authentication flow","description":"Mock the login endpoint and protect task routes with token or session.","due_date":"2025-11-08","is_completed":false,"completed_at":null,"created_at":"2025-10-30 08:50:00","updated_at":"2025-11-02 17:10:00"},{"id":5,"title":"Write documentation","description":"Document setup steps, endpoints, and deployment instructions.","due_date":"2025-11-15","is_completed":false,"completed_at":null,"created_at":"2025-11-04 13:25:00","updated_at":"2025-11-05 18:00:00"}],"meta":{"total":5,"per_page":10,"current_page":1,"last_page":1,"from":1,"to":5}}}}}},"401":{"description":"error","content":{"application\/json":{"schema":{"type":"object","example":{"message":"Unauthorized"}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X GET -G \"http:\/\/mileapp-api.test\/api\/tasks\" \\\n    -H \"Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\"\n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/mileapp-api.test\/api\/tasks\");\n\nlet params = {\n    \"page\": \"1\",\n    \"per_page\": \"10\",\n    \"filter[is_completed]\": \"\",\n    \"sort_by\": \"created_at\",\n    \"sort_order\": \"desc\",\n};\n\nObject.keys(params).forEach(key => url.searchParams.append(key, params[key]));\n\nlet headers = {\n    \"Authorization\": \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Authorization\" => \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$query = [\n    \"page\" => \"1\",\n    \"per_page\" => \"10\",\n    \"filter[is_completed]\" => \"\",\n    \"sort_by\" => \"created_at\",\n    \"sort_order\" => \"desc\",\n];\n\n\n$ch = curl_init(\"http:\/\/mileapp-api.test\/api\/tasks?\" . http_build_query($data));\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"post":{"security":[{"BearerAuth":[]}],"tags":["Tasks"],"operationId":"Create a task.","description":"","requestBody":{"description":"","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["title","description","due_date"],"properties":{"title":{"type":"string","example":"Task 6","description":"The title of the task."},"description":{"type":"string","example":"Testing tas","description":"A short description."},"due_date":{"type":"date","example":"2025-11-07","description":"The task\u2019s due date."}},"example":{"title":"Task 6","description":"Testing tas","due_date":"2025-11-07"}}}}},"parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"data":{"id":6,"title":"Task 6","description":"Testing task","due_date":"2025-11-07","is_completed":false,"completed_at":null,"created_at":"2025-11-07T12:12:45+00:00","updated_at":"2025-11-07T12:12:45+00:00"}}}}}},"401":{"description":"error","content":{"application\/json":{"schema":{"type":"object","example":{"message":"Unauthorized"}}}}},"422":{"description":"error","content":{"application\/json":{"schema":{"type":"object","example":{"message":"The given data was invalid.","errors":{"title":["The title field is required."],"description":["The description field is required."],"due_date":["The due date field is required."]}}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/mileapp-api.test\/api\/tasks\" \\\n    -H \"Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\" \\\n\n    -d \"title\"=\"Task 6\" \\\n    -d \"description\"=\"Testing tas\" \\\n    -d \"due_date\"=\"2025-11-07\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/mileapp-api.test\/api\/tasks\");\n\nlet headers = {\n    \"Authorization\": \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"title\": \"Task 6\",\n    \"description\": \"Testing tas\",\n    \"due_date\": \"2025-11-07\",\n})\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Authorization\" => \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"title\" =>  \"Task 6\",\n    \"description\" =>  \"Testing tas\",\n    \"due_date\" =>  \"2025-11-07\",\n];\n\n\n$ch = curl_init(\"http:\/\/mileapp-api.test\/api\/tasks\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_POSTFIELDS, $body);\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/tasks\/{task}":{"get":{"security":[{"BearerAuth":[]}],"tags":["Tasks"],"operationId":"Get a specific task","description":"","parameters":[{"in":"path","name":"task","description":"The ID of the task.","required":true,"schema":{"type":"integer","example":1}}],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"data":{"id":1,"title":"Design login page","description":"Create a responsive login form with validation and error handling.","due_date":"2025-11-10","is_completed":false,"completed_at":"2025-11-02 12:00:00","created_at":"2025-11-01 09:30:00","updated_at":"2025-11-03 14:45:00"}}}}}},"401":{"description":"error","content":{"application\/json":{"schema":{"type":"object","example":{"message":"Unauthorized"}}}}},"404":{"description":"error","content":{"application\/json":{"schema":{"type":"object","example":{"message":"Task not found"}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X GET -G \"http:\/\/mileapp-api.test\/api\/tasks\/{task}\" \\\n    -H \"Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\"\n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/mileapp-api.test\/api\/tasks\/{task}\");\n\nlet headers = {\n    \"Authorization\": \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Authorization\" => \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/mileapp-api.test\/api\/tasks\/{task}\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"put":{"security":[{"BearerAuth":[]}],"tags":["Tasks"],"operationId":"Update a task.","description":"","requestBody":{"description":"","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["title","description","due_date"],"properties":{"title":{"type":"string","example":"Task 6","description":"The title of the task."},"description":{"type":"string","example":"Testing tas","description":"A short description."},"due_date":{"type":"date","example":"2025-11-07","description":"The task\u2019s due date."},"is_completed":{"type":"boolean","example":true,"description":"nullable The task\u2019s status."},"completed_date":{"type":"date","example":"2025-11-07","description":"nullable The task\u2019s completed date."}},"example":{"title":"Task 6","description":"Testing tas","due_date":"2025-11-07","is_completed":true,"completed_date":"2025-11-07"}}}}},"parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"data":{"id":1,"title":"task 6","description":"lorem ipsum","due_date":"2025-11-07","is_completed":true,"completed_at":"2025-11-06","created_at":"2025-11-01T09:30:00+00:00","updated_at":"2025-11-07T12:22:29+00:00"}}}}}},"401":{"description":"error","content":{"application\/json":{"schema":{"type":"object","example":{"message":"Unauthorized"}}}}},"422":{"description":"error","content":{"application\/json":{"schema":{"type":"object","example":{"message":"The given data was invalid.","errors":{"title":["The title field is required."],"description":["The description field is required."],"due_date":["The due date field is required."]}}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X PUT \"http:\/\/mileapp-api.test\/api\/tasks\/{task}\" \\\n    -H \"Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\" \\\n\n    -d \"title\"=\"Task 6\" \\\n    -d \"description\"=\"Testing tas\" \\\n    -d \"due_date\"=\"2025-11-07\" \\\n    -d \"is_completed\"=\"1\" \\\n    -d \"completed_date\"=\"2025-11-07\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/mileapp-api.test\/api\/tasks\/{task}\");\n\nlet headers = {\n    \"Authorization\": \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"title\": \"Task 6\",\n    \"description\": \"Testing tas\",\n    \"due_date\": \"2025-11-07\",\n    \"is_completed\": \"1\",\n    \"completed_date\": \"2025-11-07\",\n})\n\nfetch(url, {\n    method: \"PUT\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Authorization\" => \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"title\" =>  \"Task 6\",\n    \"description\" =>  \"Testing tas\",\n    \"due_date\" =>  \"2025-11-07\",\n    \"is_completed\" =>  \"1\",\n    \"completed_date\" =>  \"2025-11-07\",\n];\n\n\n$ch = curl_init(\"http:\/\/mileapp-api.test\/api\/tasks\/{task}\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"PUT\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"delete":{"security":[{"BearerAuth":[]}],"tags":["Tasks"],"operationId":"Delete a specific task","description":"","parameters":[{"in":"path","name":"task","description":"The ID of the task.","required":true,"schema":{"type":"integer","example":1}}],"responses":{"204":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":[]}}}},"404":{"description":"error","content":{"application\/json":{"schema":{"type":"object","example":{"message":"Task not found"}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X DELETE \"http:\/\/mileapp-api.test\/api\/tasks\/{task}\" \\\n    -H \"Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\"\n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/mileapp-api.test\/api\/tasks\/{task}\");\n\nlet headers = {\n    \"Authorization\": \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"DELETE\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Authorization\" => \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/mileapp-api.test\/api\/tasks\/{task}\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"DELETE\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}}},"x-tagGroups":[]}